Extract Cocina mappers to app/mappers/
