+++ app/jobs/deposit_work_job.rb
@@ -52,7 +52,7 @@ class DepositWorkJob < ApplicationJob
   end
 
   def mapped_cocina_object
-    @mapped_cocina_object ||= ToCocina::Work::Mapper.call(work_form:, content:, source_id: "h3:object-#{work.id}")
+    @mapped_cocina_object ||= WorkMapper.call(work_form:, content:, source_id: "h3:object-#{work.id}")
   end
 
   def perform_persist
