+++ app/roundtrippers/work.rb
@@ -35,9 +35,9 @@ module Roundtrippers
     attr_reader :work_form, :content
 
     def roundtripped_cocina_object
-      ToCocina::Work::Mapper.call(work_form:,
-                                  content:,
-                                  source_id: normalized_original_cocina_object.identification&.sourceId)
+      WorkMapper.call(work_form:,
+                      content:,
+                      source_id: normalized_original_cocina_object.identification&.sourceId)
     end
 
     def normalized_original_cocina_object
