+++ spec/mappers/work_mapper/access_mapper_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe ToCocina::Work::AccessMapper do
+RSpec.describe WorkMapper::AccessMapper do
   subject(:access) { described_class.call(work_form:) }
 
   context 'when stanford access' do
