+++ spec/mappers/work_mapper/events_mapper_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe ToCocina::Work::EventsMapper do
+RSpec.describe WorkMapper::EventsMapper do
   describe 'publication events' do
     subject(:events) { described_class.call(work_form:) }
