+++ spec/mappers/work_mapper/identification_mapper_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe ToCocina::Work::IdentificationMapper do
+RSpec.describe WorkMapper::IdentificationMapper do
   subject(:identification) { described_class.call(work_form:, source_id:) }
 
   let(:source_id) { source_id_fixture }
