+++ spec/roundtrippers/collection_spec.rb
@@ -33,7 +33,7 @@ RSpec.describe Roundtrippers::Collection, type: :mapping do
 
   context 'when cocina raises a validation error' do
     before do
-      allow(ToCocina::Collection::Mapper).to receive(:call).and_raise(Cocina::Models::ValidationError)
+      allow(CollectionMapper).to receive(:call).and_raise(Cocina::Models::ValidationError)
     end
 
     let(:cocina_object) { collection_with_metadata_fixture }
